refactor(cli): simplify `@expo/ws-tunnel` integration
