+++ packages/@expo/cli/package.json
@@ -55,6 +55,7 @@
     "@expo/prebuild-config": "^8.0.17",
     "@expo/rudder-sdk-node": "^1.1.1",
     "@expo/spawn-async": "^1.7.2",
+    "@expo/ws-tunnel": "^1.0.1",
     "@expo/xcpretty": "^4.3.0",
     "@react-native/dev-middleware": "0.76.5",
     "@urql/core": "^5.0.6",
